backend/requirements.txt
# backend/requirements.txt
django>=4.2,<5.0
djangorestframework>=3.14,<4.0
django-cors-headers>=4.0,<5.0
djangorestframework-simplejwt>=5.3,<6.0
celery[redis]>=5.3,<6.0
redis>=4.5,<5.0
psycopg2-binary>=2.9,<3.0
boto3>=1.28,<2.0

# PyTorch with CUDA support
torch>=2.0,<3.0
torchaudio>=2.0,<3.0
numpy>=1.24,<1.26

# Audio processing
demucs>=4.0,<5.0
soundfile>=0.12,<1.0
librosa>=0.10,<1.0

# RVC Training dependencies (GPU versions)
faiss-gpu>=1.7.4,<2.0
praat-parselmouth>=0.4.3,<1.0
pyworld>=0.3.4,<1.0
scikit-learn>=1.3,<2.0
fairseq>=0.12.2,<1.0

# Misc
python-dotenv>=1.0,<2.0
pydantic>=2.0,<3.0